Feature > Network compatibility admission (#50)
## Summary

- Complete the four-message compatibility admission lifecycle with
rejection-only `D6RS`, exact `D6RO`/`D6RK`/`D6RC` identity handling,
atomic host commit, rollback, and quota release.
- Serialize Cancel, acceptance enqueue, and terminal publication through
one admission-attempt gate while preserving authoritative `receivedAt`
deadline precedence and race-safe input seal/drain behavior.
- Enforce release, protocol, capability, content-manifest, capacity,
match-state, and host-policy admission outcomes with stable
non-disclosing user-visible copy.
- Harden content-manifest traversal on Linux and Windows with pinned
roots and ancestors, identity and metadata revalidation, bounded
traversal, no-follow/reparse protections, ordinal path checks, and
fail-closed mutation handling.
- Use Win32 wide/handle traversal for DOS, extended DOS, UNC, extended
UNC, and volume paths without passing final-handle paths through
`std::filesystem` canonicalization.
- Cover compatibility, transport, cancellation/deadline races,
accounting, concrete queue ordering, process behavior, Windows sharing
semantics, hardlinks, and ordinal case replacement.
- Check out exact pull-request heads in Linux and native Windows
workflows. Native Windows runs five CTests: transport unit, transport
process, trust policy, admission compatibility, and admission process;
resolver-helper process evidence remains Linux-only.
- Integrate the reviewed UX operational-evidence manifest and supporting
implementation, test, workflow, and native build-script references.

Closes #30

## Verification

- Focused canonical Linux Release admission compatibility/process run
passed: 2/2 tests.
- Full canonical Linux Release build and CTest run passed: 13/13 tests.
- ASan/UBSan with leak detection passed 5/5 selected transport,
admission compatibility/process, and resolver-helper process tests with
no sanitizer or leak findings.
- Release stress ran transport and admission compatibility three times
each: 6/6 invocations passed.
- MinGW-w64 x86-64 Release cross-build and applicable test registration
passed for all targets.
- Win32 static/path seam validation, `actionlint`, and `git diff
--check` passed.
- The final evidence-metadata-only commit changed one line in
`docs/screenshots/README.md`; all 20 local Markdown links and 10 cited
evidence paths resolved.
- Exact-head hosted checks passed at
`79d9ef5ecdaa7a8890ee9e84fc726e3cc01b5d3e`: Linux 13/13 and native
Windows 5/5.
- Tester, review/security, visual, and hosted-readiness gates are
satisfied. Product accepted AC-001 through AC-025.

## Known risks or limitations

- This change delivers compatibility admission and its operational
scaffold; it does not yet provide authoritative replicated simulation,
reconnect state restoration, or graphical host/join/lobby flows, so it
does not by itself provide playable end-to-end network support.

## Follow-ups

- #34 — Replicate canonical match and world state with stable
identities.
- #36 — Handle reconnect and approved state restoration alongside
disconnect, shutdown, and host loss.
- #38 — Deliver graphical host, join, lobby, status, error,
cancellation, and recovery UX.

---------

+++ source/network/NetworkTrustPolicy.h
@@ -226,7 +226,7 @@ namespace Duel6::Network::Trust {
 
     using ConnectionId = std::uint64_t;
     using ParticipantId = std::uint64_t;
-    using PlayerSlotId = std::uint32_t;
+    using PlayerSlotId = std::uint64_t;
     enum class AuthorityAction { HostOnly, OwnReadiness, OwnProposal, Leave, PlayerInput };
     struct AuthorizationDecision {
         bool allowed = false;
